+++ shapiq/explainer/__init__.py
@@ -1,6 +1,6 @@
 """Explainer objects, including TreeSHAP-IQ."""
 
-from ._base import Explainer
+from .base import Explainer
 from .tabpfn import TabPFNExplainer
 from .tabular import TabularExplainer
 from .tree import TreeExplainer