+++ shapiq/utils/custom_types.py
@@ -2,7 +2,29 @@
 
 from __future__ import annotations
 
-from typing import TypeVar
+from typing import Literal, TypeVar
 
 # Model type for all machine learning models
 Model = TypeVar("Model")
+IndexType = Literal[
+    "SII",
+    "BII",
+    "CHII",
+    "Co-Moebius",
+    "SGV",
+    "BGV",
+    "CHGV",
+    "IGV",
+    "EGV",
+    "k-SII",
+    "STII",
+    "FSII",
+    "kADD-SHAP",
+    "FBII",
+    "SV",
+    "BV",
+    "JointSV",
+    "Moebius",
+    "ELC",
+    "EC",
+]